requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
b2sdk==2.10.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
deprecation==2.1.0
Django==5.2.6
django-appconf==1.1.0
django-autoslug==1.9.9
django-backblaze-b2==7.0.0
django-cleanup==9.0.0
django-filter==25.1
django-imagekit==5.0.0
django-taggit==6.1.0
django-unfold==0.65.0
djangorestframework==3.16.1
Faker==37.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
logfury==1.0.1
mysqlclient==2.2.7
packaging==25.0
pilkit==3.0
pillow==11.3.0
postgrest==2.21.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
python-dotenv==1.1.1
realtime==2.21.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.21.1
StrEnum==0.4.15
supabase==2.21.1
supabase-auth==2.21.1
supabase-functions==2.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
